Slightly refactor code
Split the bin/ folder into bin/ and lib/
lib/ shouldn't depend on bin/
