+++ bin/color_protein.py
@@ -1,4 +1,4 @@
-from utils import *
+from lib.utils import *
 
 import matplotlib
