+++ lib/sequence_parser_utils.py
@@ -1,0 +1,86 @@
+from dateutil.parser import parse as dparse
+
+def parse_viprbrc(entry):
+    fields = entry.split('|')
+    if fields[7] == 'NA':
+        date = None
+    else:
+        date = fields[7].split('/')[0]
+        date = dparse(date.replace('_', '-'))
+
+    country = fields[9]
+    from locations import country2continent
+    if country in country2continent:
+        continent = country2continent[country]
+    else:
+        country = 'NA'
+        continent = 'NA'
+
+    from mammals import species2group
+
+    meta = {
+        'strain': fields[5],
+        'host': fields[8],
+        'group': species2group[fields[8]],
+        'country': country,
+        'continent': continent,
+        'dataset': 'viprbrc',
+    }
+    return meta
+
+def parse_nih(entry):
+    fields = entry.split('|')
+
+    country = fields[3]
+    from locations import country2continent
+    if country in country2continent:
+        continent = country2continent[country]
+    else:
+        country = 'NA'
+        continent = 'NA'
+
+    meta = {
+        'strain': 'SARS-CoV-2',
+        'host': 'human',
+        'group': 'human',
+        'country': country,
+        'continent': continent,
+        'dataset': 'nih',
+    }
+    return meta
+
+def parse_gisaid(entry):
+    fields = entry.split('|')
+
+    type_id = fields[1].split('/')[1]
+
+    if type_id in { 'bat', 'canine', 'cat', 'env', 'mink',
+                    'pangolin', 'tiger' }:
+        host = type_id
+        country = 'NA'
+        continent = 'NA'
+    else:
+        host = 'human'
+        from locations import country2continent
+        if type_id in country2continent:
+            country = type_id
+            continent = country2continent[country]
+        else:
+            country = 'NA'
+            continent = 'NA'
+
+    from mammals import species2group
+
+    meta = {
+        'strain': fields[1],
+        'host': host,
+        'group': species2group[host].lower(),
+        'country': country,
+        'continent': continent,
+        'dataset': 'gisaid',
+    }
+    return meta
+
+
+def parse_aggregated_mut_escapes(entry):
+   ...
